pio/requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.0.12
Django==3.2.13
django-filter==21.1
djangorestframework==3.13.1
docopt==0.6.2
idna==3.3
importlib-metadata==4.11.4
pytz==2022.1
requests==2.28.0
sqlparse==0.4.2
typing_extensions==4.2.0
urllib3==1.26.9
yarg==0.1.9
zipp==3.8.0
